requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.4.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
commonmark==0.9.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.0.12
fire==0.4.0
fsspec==2022.2.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.7.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
lazy-object-proxy==1.4.3
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.0.0
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pylint==2.6.0
PyMuPDF==1.19.6
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rich==12.2.0
Send2Trash==1.8.0
simplify-docx==0.1.1
six==1.15.0
soupsieve==2.3.1
stack-data==0.2.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
virtualenv==20.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.12.1
